extract query logic to useComunicaQuery hook
